pyproject.toml
[tool.poetry]
name = "fastapi-ai-backend"
version = "0.1.0"
description = "Diet App Backend API with FastAPI and AI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.30"
pydantic = "^2.10.6"
python-jose = "^3.3.0"
python-multipart = "^0.0.18"
bcrypt = "^4.1.2"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
google-cloud-firestore = "^2.20.0"
pandas = "*"
scikit-learn = "*"
joblib = "*"
PyJWT = "^2.8.0"
pydantic-settings = "^2.9.1"
aiofiles = "^24.1.0"
boto3 = "1.26.76"
botocore = "1.29.76"
annotated-types = "^0.7.0"
anyio = "^4.8.0"
cachetools = "^5.5.1"
certifi = "^2025.1.31"
charset-normalizer = "^3.4.1"
click = "^8.1.8"
colorama = "^0.4.6"
google-api-core = "^2.24.1"
google-auth = "^2.38.0"
google-cloud-core = "^2.4.1"
googleapis-common-protos = "^1.67.0"
grpcio = "^1.70.0"
grpcio-status = "^1.70.0"
h11 = "^0.14.0"
idna = "^3.10"
proto-plus = "^1.26.0"
protobuf = "^5.29.3"
pyasn1 = "^0.6.1"
pyasn1_modules = "^0.4.1"
pydantic_core = "^2.27.2"
requests = "^2.32.3"
rsa = "^4.9"
sniffio = "^1.3.1"
starlette = "^0.45.3"
typing_extensions = "^4.12.2"
urllib3 = "1.26.18"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"